Add different trained models
Created two folders. One "Models" for the working models on each step. It has the alpha model, and the beta model used with the old data and with the new data for track1 (both working with speed limit 10)
The other folder "Test_models" contains model that were trained and tested but failed to run the car.