Add more gamma models
To improve the performance multiple models are trained by varying the batch sizes and steps and thus the epochs.
The increase of batches size has a negative effect.
The increase of steps improve the performance on track2 for detecting the correct lanes, but the car had trouble driving on the bridge on track1.