Add gamma model
Increasing the steps help to improve the performance on track2. It is able to detect the lanes on low speed limit (max. 8) but shows some trouble on driving up the hills and sharp curves.
On track1 shows also improvement driving the bridge but tends to drive near the line instead in the middle.